+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `RocketIcon` and `RocketCircleIcon` component.
+- Feature: Add `RocketIcon` and `RocketCircleIcon` components.
 - Feature: Create `ProgressRing` component.
 
 ## [2.64.0] - 2020-04-07